+++ test/pls.jl
@@ -181,6 +181,7 @@ end
     refit!(fm, float(MixedModels.dataset(:dyestuff2)[:yield]); progress=false) # restore the model in the cache
 end
 
+
 @testset "penicillin" begin
     fm = only(models(:penicillin))
     @test size(fm) == (144, 1, 30, 2)
@@ -374,6 +375,12 @@ end
     @test keys(b3tbl) == (:subj,)
     @test isa(b3tbl, NamedTuple)
     @test Tables.istable(only(b3tbl))
+
+    @testset "PosDefException from constant response" begin
+        slp = MixedModels.dataset(:sleepstudy)
+        @test_throws ArgumentError("The response is constant and thus model fitting has failed") refit!(fm, zero(slp.reaction); progress=false)
+        refit!(fm, slp.reaction; progress=false)
+    end
 
     simulate!(fm)  # to test one of the unscaledre methods
     # must restore state of fm as it is cached in the global fittedmodels